# Fix a crash when deserializing empty sparseList