--- a/bitoperations_test.go
+++ b/bitoperations_test.go
@@ -1,8 +1,9 @@
 package gohll
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestSliceUint32(t *testing.T) {