Discussion (old): https://groups.google.com/forum/#!topic/golang-nuts/VnFs2Cv0_UY
Snippet reproducing the problem: 
https://play.golang.org/p/E_EAipsfR7

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/mynameisfiber/gohll/6)
<!-- Reviewable:end -->