@mynameisfiber 2015-12-16
Could you run the speed benchmark before/after these commits?