@sasha-s 2015-12-16
new code:
go test -bench .
PASS
BenchmarkAddSparse-8        	 1000000	      2005 ns/op	     109 B/op	       3 allocs/op
BenchmarkAddNormal-8        	 5000000	       251 ns/op	      32 B/op	       2 allocs/op
BenchmarkAddSparseFNV1A-8   	 1000000	      1883 ns/op	     125 B/op	       4 allocs/op
BenchmarkAddNormalFNV1A-8   	 5000000	       258 ns/op	      48 B/op	       3 allocs/op
BenchmarkCardinalityNormal-8	     300	   4545612 ns/op	       0 B/op	       0 allocs/op
BenchmarkCardinalitySparse-8	    2000	    761730 ns/op	      73 B/op	       0 allocs/op
ok  	github.com/sasha-s/gohll	12.357s

old code:
go test -bench .
PASS
BenchmarkAddSparse-8        	 1000000	      2012 ns/op	     109 B/op	       3 allocs/op
BenchmarkAddNormal-8        	10000000	       247 ns/op	      32 B/op	       2 allocs/op
BenchmarkAddSparseFNV1A-8   	 1000000	      1839 ns/op	     125 B/op	       4 allocs/op
BenchmarkAddNormalFNV1A-8   	 5000000	       257 ns/op	      48 B/op	       3 allocs/op
BenchmarkCardinalityNormal-8	     300	   4570232 ns/op	       0 B/op	       0 allocs/op
BenchmarkCardinalitySparse-8	    2000	    757497 ns/op	      73 B/op	       0 allocs/op
ok  	github.com/mynameisfiber/gohll	13.516s

---

Review status: 0 of 5 files reviewed at latest revision, all discussions resolved.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/mynameisfiber/gohll/6#-:-K5gDSZo1S3R07JRrU5z:564312249)*
<!-- Sent from Reviewable.io -->